refactor: changed environment files
