+++ Breaking/app/src/main/java/com/dope/breaking/model/request/RequestGoogleToken.kt
@@ -1,3 +1,3 @@
-package com.dope.breaking.model
+package com.dope.breaking.model.request
 
 class RequestGoogleToken(val idToken: String, val accessToken: String)
