+++ Breaking/app/src/main/java/com/dope/breaking/model/request/RequestSignUp.kt
@@ -1,4 +1,4 @@
-package com.dope.breaking.model
+package com.dope.breaking.model.request
 
 import org.json.JSONObject
